Reorganized install.pl
No funtional change. Just some reorganizations.

+++ scripts/install.pl
@@ -4,36 +4,6 @@ use strict;
 use warnings;
 use File::Basename;
 
-my $moddir = shift @ARGV;
-my $mode = shift @ARGV;
-my $dir = shift @ARGV;
-
-die "No MODULE_DIR" unless defined $moddir;
-die "No mode" unless defined $mode;
-die "No INSTALL_DIR" unless defined $dir;
-
-my @files = @ARGV;
-my $files = join " ", @files;
-
-die "No files to install" unless @files;
-
-if (!-d $dir) {
-    system("mkdir -p $dir") == 0
-        or die "Invalid directory $dir";
-}
-
-my @libscripts;
-push @libscripts, "USBScripts";
-if ($dir =~ m/haps_commands/) {
-    push @libscripts, "HAPSScripts";
-}
-
-my $script;
-my $modules;
-for $script (@libscripts) {
-    $modules .= "use $script;\n";
-}
-
 sub read_file {
     my $file = shift;
 
@@ -54,33 +24,73 @@ sub write_file {
     close $fh;
 }
 
-my $file;
-for $file (@files) {
-    if (!-f $file) {
-        die "Invalid file $file";
+sub tag_use_module {
+    my $dir = shift;
+
+    my @libscripts;
+    push @libscripts, "USBScripts";
+    if ($dir =~ m/haps_commands/) {
+        push @libscripts, "HAPSScripts";
     }
 
-    my $dest = "$dir/" . basename($file);
-
-    my $perl = 0;
-    my $contents = read_file($file);
-    if ($contents =~ m/(^\#\!\/usr\/bin\/perl.*\n)/) {
-        $perl = 1;
-        $contents =~ s/($1)/$1\nuse lib '$moddir';\n$modules/;
-        $file = "$file.install";
-        write_file($file, $contents);
+    my $modules;
+    for (@libscripts) {
+        $modules .= "use $_;\n";
     }
 
-    my $cmd = "install -m $mode -D $file $dest";
-    print "$cmd\n";
+    return $modules;
+}
 
-    my $ret = system($cmd);
-    if ($perl) {
-        unlink($file);
+sub main {
+    my $moddir = shift @ARGV;
+    my $mode = shift @ARGV;
+    my $dir = shift @ARGV;
+
+    die "No MODULE_DIR" unless defined $moddir;
+    die "No mode" unless defined $mode;
+    die "No INSTALL_DIR" unless defined $dir;
+
+    my @files = @ARGV;
+    my $files = join " ", @files;
+
+    die "No files to install" unless @files;
+
+    if (!-d $dir) {
+        system("mkdir -p $dir") == 0
+            or die "Invalid directory $dir";
     }
 
-    if ($ret) {
-        die "Couldn't install $file";
+    my $modules = tag_use_module($dir);
+
+    my $file;
+    for $file (@files) {
+        if (!-f $file) {
+            die "Invalid file $file";
+        }
+
+        my $dest = "$dir/" . basename($file);
+
+        my $perl = 0;
+        my $contents = read_file($file);
+        if ($contents =~ m/(^\#\!\/usr\/bin\/perl.*\n)/) {
+            $perl = 1;
+            $contents =~ s/($1)/$1\nuse lib '$moddir';\n$modules/;
+            $file = "$file.install";
+            write_file($file, $contents);
+        }
+
+        my $cmd = "install -m $mode -D $file $dest";
+        print "$cmd\n";
+
+        my $ret = system($cmd);
+        if ($perl) {
+            unlink($file);
+        }
+
+        if ($ret) {
+            die "Couldn't install $file";
+        }
     }
 }
 
+main();
